+++ src/app/ledger/page.tsx
@@ -6,6 +6,8 @@ import {
   Box,
   CircularProgress,
 } from '@mui/material';
+import { useTheme } from '@mui/material/styles';
+import useMediaQuery from '@mui/material/useMediaQuery';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { DateCalendar } from '@mui/x-date-pickers/DateCalendar';
@@ -16,6 +18,7 @@ import BackgroundWrapper from '@/app/components/BackgroundWrapper';
 import { SummaryBox } from './components/SummaryBox';
 import { AddPanel } from './components/AddPanel';
 import { RecordHistory } from './components/History';
+import DesktopView from './views/Desktop';
 
 function LedgerCalendar(
   {date, handler}: {date: Date, handler: (date: Date) => void}
@@ -32,66 +35,17 @@ function LedgerCalendar(
 }
 
 function LedgerPage() {
-  const router = useRouter();
-  const { isLoggedIn } = useAuth();
-  const { records, loading, error, addRecord, refreshRecords } = useRecords();
-  const today = new Date();
-  const [selectedDate, setSelectedDate] = useState<Date>(today);
+  const theme = useTheme();
+  const small = useMediaQuery(theme.breakpoints.down('md'));
 
-  useEffect(() => {
-    if (!isLoggedIn) {
-      router.replace('/login');
-    }
-  }, [isLoggedIn, router]);
-
-  // TODO: need to handle null?
-  const handleDateChange = (newDate: Dayjs) => {
-    if (newDate && newDate.isValid()) {
-      setSelectedDate(newDate.toDate());
-    }
-  }
-
-  if (loading) {
+  if (small) {
     return (
-      <BackgroundWrapper>
-        <Box
-          display="flex"
-          justifyContent="center"
-          alignItems="center"
-          minHeight="100vh"
-        >
-          <CircularProgress />
-        </Box>
-      </BackgroundWrapper>
-    );
+      <div>Working on this</div>
+    )
   }
 
-  return (
-    <MainLayout>
-      <div className="ledger-layout">
-        {/* Column one: Calender + Summary + AddPanel */}
-        <div className="ledger-left-column">
-          <div className="flex-1 flex justify-between items-start ledger-block-border">
-            <LedgerCalendar
-              date={selectedDate}
-              handler={handleDateChange}
-            />
-            <SummaryBox date={selectedDate}/>
-          </div>
-          <AddPanel
-            selectedDate={selectedDate}
-            onAddRecord={addRecord}
-            onRefresh={refreshRecords}
-          />
-        </div>
-
-        {/* Column two: History */}
-        <div className="ledger-right-column ledger-block-border">
-          <RecordHistory records={records} />
-        </div>
-      </div>
-    </MainLayout>
-  );
+  // wide screen by default
+  return <DesktopView />
 };
 
 export default LedgerPage;
